refactor(anchor): move instruction scripts into an instructions folder
Why:
* Separate write flows from view scripts and program helpers for a clearer
  per-variant layout that native and kit will mirror.

What changed:
* Move deposit, request-redeem, instant-redeem and claim-withdrawal into
  anchor/instructions and fix their relative imports.

Impact:
* None. File moves and import paths only.
