+++ packages/trubill-vault/anchor/instructions/request-redeem.ts
@@ -2,13 +2,13 @@ import { fileURLToPath } from "node:url";
 import { BN, type Program } from "@coral-xyz/anchor";
 import { getAssociatedTokenAddressSync, TOKEN_2022_PROGRAM_ID } from "@solana/spl-token";
 import { type PublicKey, SystemProgram } from "@solana/web3.js";
-import { trubill } from "../common/amounts";
-import { type TrubillVault } from "../common/idls/trubill_vault";
-import { getWalletKeypair } from "../common/web3/env";
-import * as pda from "../common/web3/pda";
-import { buildSignAndProcessTxV0 } from "../common/web3/tx";
-import { getLatestCompletedEpoch } from "./epoch";
-import { getProvider, getTrubillVaultProgram } from "./program";
+import { trubill } from "../../common/amounts";
+import { type TrubillVault } from "../../common/idls/trubill_vault";
+import { getWalletKeypair } from "../../common/web3/env";
+import * as pda from "../../common/web3/pda";
+import { buildSignAndProcessTxV0 } from "../../common/web3/tx";
+import { getLatestCompletedEpoch } from "../epoch";
+import { getProvider, getTrubillVaultProgram } from "../program";
 
 /** Build a request-redeem instruction: burn TruBILL now and record a claim payable after settlement. */
 export async function requestRedeemIx(params: {
